fix: remove sass and enable server components
