--- .idea/workspace.xml
@@ -1,138 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="AutoImportSettings">
-    <option name="autoReloadType" value="SELECTIVE" />
-  </component>
-  <component name="ChangeListManager">
-    <list default="true" id="64a5bab2-6838-4d52-b567-9f6494c33e81" name="Changes" comment="">
-      <change afterPath="$PROJECT_DIR$/apps/nuxt-showcase-app/.nvmrc" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/apps/nuxt-showcase-app/assets/styles/app.css" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/apps/nuxt-showcase-app/assets/styles/main.css" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/apps/nuxt-showcase-app/assets/styles/normalize.css" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/apps/nuxt-showcase-app/assets/styles/variables.css" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/apps/nuxt-showcase-app/components/AppFilters.vue" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/apps/nuxt-showcase-app/components/AppFooter.vue" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/apps/nuxt-showcase-app/components/AppHeader.vue" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/apps/nuxt-showcase-app/components/StoreTile.vue" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/apps/nuxt-showcase-app/types/RestaurantesResponse.ts" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/apps/nuxt-showcase-app/app.vue" beforeDir="false" afterPath="$PROJECT_DIR$/apps/nuxt-showcase-app/app.vue" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/apps/nuxt-showcase-app/layouts/default.vue" beforeDir="false" afterPath="$PROJECT_DIR$/apps/nuxt-showcase-app/layouts/default.vue" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/apps/nuxt-showcase-app/package-lock.json" beforeDir="false" afterPath="$PROJECT_DIR$/apps/nuxt-showcase-app/package-lock.json" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/apps/nuxt-showcase-app/package.json" beforeDir="false" afterPath="$PROJECT_DIR$/apps/nuxt-showcase-app/package.json" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/apps/nuxt-showcase-app/pages/index.vue" beforeDir="false" afterPath="$PROJECT_DIR$/apps/nuxt-showcase-app/pages/index.vue" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/apps/nuxt-showcase-app/pages/portfolio/index.vue" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/mock-server/public/css/variables.css" beforeDir="false" afterPath="$PROJECT_DIR$/mock-server/public/css/variables.css" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/mock-server/public/customer-web.html" beforeDir="false" afterPath="$PROJECT_DIR$/mock-server/public/customer-web.html" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/package.json" beforeDir="false" afterPath="$PROJECT_DIR$/package.json" afterDir="false" />
-    </list>
-    <option name="SHOW_DIALOG" value="false" />
-    <option name="HIGHLIGHT_CONFLICTS" value="true" />
-    <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
-    <option name="LAST_RESOLUTION" value="IGNORE" />
-  </component>
-  <component name="ComposerSettings">
-    <execution />
-  </component>
-  <component name="FileTemplateManagerImpl">
-    <option name="RECENT_TEMPLATES">
-      <list>
-        <option value="Vue Composition API Component" />
-        <option value="TypeScript File" />
-      </list>
-    </option>
-  </component>
-  <component name="Git.Settings">
-    <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
-  </component>
-  <component name="MarkdownSettingsMigration">
-    <option name="stateVersion" value="1" />
-  </component>
-  <component name="ProjectId" id="2UO4peDx4vxntimA1OkcJSIs5Ir" />
-  <component name="ProjectViewState">
-    <option name="hideEmptyMiddlePackages" value="true" />
-    <option name="showLibraryContents" value="true" />
-  </component>
-  <component name="PropertiesComponent"><![CDATA[{
-  "keyToString": {
-    "RunOnceActivity.ShowReadmeOnStart": "true",
-    "WebServerToolWindowFactoryState": "false",
-    "git-widget-placeholder": "vd3v/CCEV2-2777/nuxtjs-version-of-the-page",
-    "last_opened_file_path": "/Users/viktor.borisov/Code/two-of-one/apps/nuxt-showcase-app/assets/styles",
-    "node.js.detected.package.eslint": "true",
-    "node.js.detected.package.tslint": "true",
-    "node.js.selected.package.eslint": "(autodetect)",
-    "node.js.selected.package.tslint": "(autodetect)",
-    "nodejs_package_manager_path": "npm",
-    "ts.external.directory.path": "/Users/viktor.borisov/Library/Application Support/JetBrains/Toolbox/apps/PhpStorm/ch-0/231.9161.47/PhpStorm.app/Contents/plugins/javascript-impl/jsLanguageServicesImpl/external",
-    "vue.rearranger.settings.migration": "true"
-  },
-  "keyToStringList": {
-    "com.intellij.ide.scratch.ScratchImplUtil$2/New Scratch File": [
-      "JSON"
-    ],
-    "vue.recent.templates": [
-      "Vue Composition API Component"
-    ]
-  }
-}]]></component>
-  <component name="RecentsManager">
-    <key name="CopyFile.RECENT_KEYS">
-      <recent name="$PROJECT_DIR$/apps/nuxt-showcase-app/assets/styles" />
-    </key>
-  </component>
-  <component name="RunManager" selected="npm.start">
-    <configuration name="build" type="js.build_tools.npm" temporary="true" nameIsGenerated="true">
-      <package-json value="$PROJECT_DIR$/apps/nuxt-showcase-app/package.json" />
-      <command value="run" />
-      <scripts>
-        <script value="build" />
-      </scripts>
-      <node-interpreter value="project" />
-      <envs />
-      <method v="2" />
-    </configuration>
-    <configuration name="dev" type="js.build_tools.npm" temporary="true" nameIsGenerated="true">
-      <package-json value="$PROJECT_DIR$/apps/nuxt-showcase-app/package.json" />
-      <command value="run" />
-      <scripts>
-        <script value="dev" />
-      </scripts>
-      <node-interpreter value="project" />
-      <envs />
-      <method v="2" />
-    </configuration>
-    <configuration name="start" type="js.build_tools.npm" temporary="true" nameIsGenerated="true">
-      <package-json value="$PROJECT_DIR$/apps/nuxt-showcase-app/package.json" />
-      <command value="run" />
-      <scripts>
-        <script value="start" />
-      </scripts>
-      <node-interpreter value="project" />
-      <envs />
-      <method v="2" />
-    </configuration>
-    <recent_temporary>
-      <list>
-        <item itemvalue="npm.start" />
-        <item itemvalue="npm.build" />
-        <item itemvalue="npm.dev" />
-      </list>
-    </recent_temporary>
-  </component>
-  <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
-  <component name="TaskManager">
-    <task active="true" id="Default" summary="Default task">
-      <changelist id="64a5bab2-6838-4d52-b567-9f6494c33e81" name="Changes" comment="" />
-      <created>1692797477219</created>
-      <option name="number" value="Default" />
-      <option name="presentableId" value="Default" />
-      <updated>1692797477219</updated>
-      <workItem from="1692797478600" duration="729000" />
-      <workItem from="1692798498147" duration="7662000" />
-    </task>
-    <servers />
-  </component>
-  <component name="TypeScriptGeneratedFilesManager">
-    <option name="version" value="3" />
-  </component>
-</project>
