+++ apps/nuxt-showcase-app/package.json
@@ -18,8 +18,5 @@
   },
   "engines": {
     "node": "16.x || 18.x"
-  },
-  "dependencies": {
-    "sass": "^1.66.1"
   }
 }
